Use double precision for generated hit position

+++ src/mag_field_utils.h
@@ -218,7 +218,7 @@ class MagFieldUtils
   MagFieldUtils(const tpcrs::Configurator& cfg, const CoordTransform& trans, const MagField::EBField map, const float factor, int mode);
 
   void    B3DFieldTpc ( const float xTpc[], float BTpc[], int Sector = -1 );
-  void    BFieldTpc ( const float xTpc[], float BTpc[], int Sector = -1 );
+  void    BFieldTpc ( const double xTpc[], float BTpc[], int Sector = -1 );
 
   void    DoDistortion ( const float x[], float Xprime[], int Sector = -1 ) ;
   void    UndoDistortion ( const float x[], float Xprime[], int Sector = -1 ) ;
